+++ CMakeLists.txt
@@ -3,9 +3,8 @@ project(pypebble)
 
 set(CMAKE_CXX_STANDARD 17)
 
-# Find python and Boost - both are required dependencies
+# Find python and pybind11 - both are required dependencies
 find_package(Python 3.7 REQUIRED COMPONENTS Interpreter Development)
-find_package(Boost COMPONENTS python REQUIRED)
 find_package(pybind11 CONFIG)
 
 
@@ -37,44 +36,11 @@ add_dependencies(pypebble gen_libpebble)
 # Set up the libraries and header search paths for this target
 target_link_libraries(pypebble
         ${CMAKE_BINARY_DIR}/libpebble.so
-        ${Boost_LIBRARIES}
+        pybind11::module
         ${Python_LIBRARIES})
 target_include_directories(pypebble PRIVATE
         ${CMAKE_SOURCE_DIR}
         ${Python_INCLUDE_DIRS})
-
-# Pybind11 Version
-#pybind11_add_module(pypebble11 MODULE
-#        ${CMAKE_BINARY_DIR}/libpebble.so
-#        pebblepp/common.cpp
-#        pebblepp/iterator.cpp
-#        pebblepp/options.cpp
-#        pebblepp/pebble.cpp
-#        pebblepp/pypebble11.cpp)
-#target_link_libraries(pypebble11
-#        ${CMAKE_BINARY_DIR}/libpebble.so)
-#target_include_directories(pypebble11 PRIVATE
-#        ${CMAKE_SOURCE_DIR})
-
-add_library(pypebble11 MODULE
-        pebblepp/common.cpp
-        pebblepp/iterator.cpp
-        pebblepp/options.cpp
-        pebblepp/pebble.cpp
-        pebblepp/pypebble11.cpp)
-target_link_libraries(pypebble11
-        ${CMAKE_BINARY_DIR}/libpebble.so
-        pybind11::module
-        ${Python_LIBRARIES})
-target_include_directories(pypebble11 PRIVATE
-        ${CMAKE_SOURCE_DIR}
-        ${Python_INCLUDE_DIRS})
-add_dependencies(pypebble11 gen_libpebble)
-#set_target_properties(pypebble11 PROPERTIES
-#        INTERPROCEDURAL_OPTIMIZATION ON
-#        CXX_VISIBILITY_PRESET ON
-#        VISIBILITY_INLINES_HIDDEN ON)
-
 
 # pebtest.out
 add_executable(pebtest.out
