+++ pebblepp/pypebble.cpp
@@ -1,63 +1,83 @@
-#include <boost/python.hpp>
-
 #include "pebblepp/iterator.h"
 #include "pebblepp/options.h"
 #include "pebblepp/pebble.h"
-using namespace boost::python;
+
+#include <pybind11/pybind11.h>
+namespace py = pybind11;
 
 namespace cockroachdb::pebble {
 
-BOOST_PYTHON_FUNCTION_OVERLOADS(basic_options_overloads, BasicOptions, 0, 1)
-BOOST_PYTHON_FUNCTION_OVERLOADS(cockroach_options_overloads, CockroachDefaultOptions, 0, 1)
-
-BOOST_PYTHON_MEMBER_FUNCTION_OVERLOADS(set_overloads, Set, 2, 3)
-BOOST_PYTHON_MEMBER_FUNCTION_OVERLOADS(delete_overloads, Delete, 1, 2)
-BOOST_PYTHON_MEMBER_FUNCTION_OVERLOADS(single_delete_overloads, SingleDelete, 1, 2)
-BOOST_PYTHON_MEMBER_FUNCTION_OVERLOADS(delete_range_overloads, DeleteRange, 2, 3)
-BOOST_PYTHON_MEMBER_FUNCTION_OVERLOADS(merge_overloads, Merge, 2, 3)
-
-DB* (*open1)(const std::string&) = &DB::Open;
-DB* (*open2)(const std::string&, const Options*) = &DB::Open;
+DB* (*dbOpen)(const std::string&) = &DB::Open;
+DB* (*dbOpenWithOpts)(const std::string&, const Options*) = &DB::Open;
 
 template <typename Copyable>
-const Copyable copyObject(const Copyable& v) {
+const Copyable CopyObject(const Copyable& v) {
   return v;
 }
 
-BOOST_PYTHON_MODULE(pypebble) {
-  class_<Options, boost::noncopyable>("Options", no_init);
-  def("BasicOptions", &BasicOptions,
-      basic_options_overloads(args("read_write"),
-                              "basic pebble options")[return_value_policy<manage_new_object>()]);
-  def("CockroachDefaultOptions", &CockroachDefaultOptions,
-      cockroach_options_overloads(
-          args("read_write"),
-          "cockroach default options")[return_value_policy<manage_new_object>()]);
-  // TODO(sarkesian): Convert the below arguments to kwargs, or to a mutable object.
-  def("PebbleOptions", &PebbleOptions,
-      args("read_write", "use_cockroach_interfaces", "l0_compaction_threshold",
-           "l0_stop_writes_threshold", "lbase_max_bytes", "levels", "max_concurrent_compactions",
-           "mem_table_size", "mem_table_stop_writes_threshold", "block_size", "index_block_size"),
-      return_value_policy<manage_new_object>(), "custom Pebble options");
+template<class T>
+T* NoInit() {
+  throw std::runtime_error("This class cannot be instantiated from Python");
+  return (T*)nullptr;
+}
 
-  enum_<IterKeyType>("IterKeyType")
+PYBIND11_MODULE(pypebble, m) {
+  // -- Module-level Functions and Utilities --
+  m.def("CGoHandles", &CGoHandles,
+        "number of live handles to Go-managed memory available to C/Python");
+  m.def("PrettyPrintKey", &PrettyPrintKey, "pretty-print a CockroachDB key");
+  m.def("PrettyScanKey", &PrettyScanKey, "scan a human-readable CockroachDB key");
+
+  m.def("Open", dbOpen, "opens a pebble DB",
+        py::arg("path"));
+  m.def("Open", dbOpenWithOpts, "opens a pebble DB",
+        py::arg("path"),
+        py::arg("options"));
+
+  // -- pebble.Options and utils --
+  py::class_<Options>(m, "Options")
+      .def(py::init(&NoInit<Options>));
+  m.def("BasicOptions", &BasicOptions,
+        "basic pebble options",
+        py::arg("read_write") = false);
+  m.def("CockroachDefaultOptions", &CockroachDefaultOptions,
+        "default options for cockroachdb pebble instances",
+        py::arg("read_write") = false);
+  m.def("PebbleOptions", &PebbleOptions,
+        "customizable pebble options",
+        py::arg("read_write"),
+        py::arg("use_cockroach_interfaces"),
+        py::arg("l0_compaction_threshold"),
+        py::arg("l0_stop_writes_threshold"),
+        py::arg("lbase_max_bytes"),
+        py::arg("levels"),
+        py::arg("max_concurrent_compactions"),
+        py::arg("mem_table_size"),
+        py::arg("mem_table_stop_writes_threshold"),
+        py::arg("block_size"),
+        py::arg("index_block_size"));
+
+  // -- pebble.Iterator and utils --
+  py::enum_<IterKeyType>(m, "IterKeyType")
       .value("POINTS_ONLY", kPointsOnly)
       .value("RANGES_ONLY", kRangesOnly)
       .value("POINTS_AND_RANGES", kPointsAndRanges);
 
-  class_<IterOptions>("IterOptions")
-      .def(init<const IterOptions&>())
-      .def("__copy__", &copyObject<IterOptions>)
-      .add_property("lower_bound", &IterOptions::LowerBound, &IterOptions::SetLowerBound)
-      .add_property("upper_bound", &IterOptions::UpperBound, &IterOptions::SetUpperBound)
-      .add_property("key_types", &IterOptions::KeyTypes, &IterOptions::SetKeyTypes);
+  py::class_<IterOptions>(m, "IterOptions")
+      .def(py::init())
+      .def(py::init<const IterOptions&>())
+      .def("__copy__", &CopyObject<IterOptions>)
+      .def_property("lower_bound", &IterOptions::LowerBound, &IterOptions::SetLowerBound)
+      .def_property("upper_bound", &IterOptions::UpperBound, &IterOptions::SetUpperBound)
+      .def_property("key_types", &IterOptions::KeyTypes, &IterOptions::SetKeyTypes);
 
-  enum_<IterValidityState>("IterValidityState")
+  py::enum_<IterValidityState>(m, "IterValidityState")
       .value("EXHAUSTED", kExhausted)
       .value("VALID", kValid)
       .value("AT_LIMIT", kAtLimit);
 
-  class_<Iterator, boost::noncopyable>("Iterator", no_init)
+  py::class_<Iterator>(m, "Iterator")
+      .def(py::init(&NoInit<Iterator>))
       //.def("Clone", &Iterator::Clone)
       .def("SeekGE", &Iterator::SeekGE)
       .def("SeekGEWithLimit", &Iterator::SeekGEWithLimit)
@@ -87,28 +107,34 @@ BOOST_PYTHON_MODULE(pypebble) {
       .def("ResetStats", &Iterator::ResetStats);
   //.def("Stats", &Iterator::Stats);
 
-  class_<DB, boost::noncopyable>("DB", no_init)
+  // -- pebble.DB --
+  py::class_<DB>(m, "DB")
+      .def(py::init(&NoInit<DB>))
       .def("Close", &DB::Close)
       .def("NumFiles", &DB::NumFiles)
-      .def("Get", &DB::Get, args("key"), "gets value of a key")
-      .def("Set", &DB::Set, set_overloads(args("key", "val", "sync"), "sets a key to a value"))
-      .def("Delete", &DB::Delete, delete_overloads(args("key", "sync"), "deletes a key"))
-      .def("SingleDelete", &DB::SingleDelete,
-           single_delete_overloads(args("key", "sync"), "single deletes a key"))
-      .def("DeleteRange", &DB::DeleteRange,
-           delete_range_overloads(args("start_key", "end_key", "sync"), "delete a range of keys"))
-      .def("Merge", &DB::Merge,
-           merge_overloads(args("key", "val", "sync"), "merge a value into a key"))
-      .def("NewIter", &DB::NewIter, return_value_policy<manage_new_object>());
-
-  def("Open", open1, args("path"), return_value_policy<manage_new_object>(),
-      "opens a new pebble.DB");
-  def("Open", open2, args("path", "options"), return_value_policy<manage_new_object>(),
-      "opens a new pebble.DB");
-  def("CGoHandles", &CGoHandles,
-      "number of live handles to Go-managed memory available to C/Python");
-  def("PrettyPrintKey", &PrettyPrintKey, "pretty-print a CockroachDB key");
-  def("PrettyScanKey", &PrettyScanKey, "scan a human-readable CockroachDB key");
+      .def("Get", &DB::Get,
+           "gets value of a key",
+           py::arg("key"))
+      .def("Set", &DB::Set, "stores a value at a key",
+           py::arg("key"),
+           py::arg("val"),
+           py::arg("sync") = true)
+      .def("Delete", &DB::Delete, "deletes a key",
+           py::arg("key"),
+           py::arg("sync") = true)
+      .def("SingleDelete", &DB::SingleDelete, "single-deletes a key",
+           py::arg("key"),
+           py::arg("sync") = true)
+      .def("DeleteRange", &DB::DeleteRange, "delete a range of keys",
+           py::arg("start_key"),
+           py::arg("end_key"),
+           py::arg("sync") = true)
+      .def("Merge", &DB::Merge, "merge a value at a key",
+           py::arg("key"),
+           py::arg("val"),
+           py::arg("sync") = true)
+      .def("NewIter", &DB::NewIter, "create a new pebble Iterator",
+           py::arg("opts"));
 }
 
 }  // namespace cockroachdb::pebble
